Support arbitrary axes for ReinterpretArray
This allows one to reinterpret arrays where indexing does not start at 1.
Axis 1 of the reinterpreted array starts at the same index as the parent.
Consequently, if the parent axis 1 is f:l, then the reinterpreted array
has axis f:(l-f+1)/r+f-1, where r is the ratio of sizes of the eltypes.